refactor: 어드민 관련 클래스 모듈 변경
