+++ maeilmail-admin/src/main/java/maeilmail/admin/AdminNoticeRequest.java
