+++ maeilmail-admin/src/main/java/maeilmail/admin/AdminNoticeTestRequest.java
